Rename logger Step2 files to Step1
